svg/painting/parsing/shape-rendering-computed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: getComputedStyle().shapeRendering</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#ShapeRenderingProperty" rel="help"/><h:meta content="shape-rendering computed value is as specified." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/computed-testcommon.js"/><script><![CDATA[

test_computed_value("shape-rendering", "auto");
test_computed_value("shape-rendering", "optimizespeed");
test_computed_value("shape-rendering", "crispedges");
test_computed_value("shape-rendering", "geometricprecision");

  ]]></script></svg>
